requirements.txt
pyttsx3==2.90
speech_recognition==3.10.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
pyaudio==0.2.13
wikipedia==1.4.0
playsound==1.3.0
pywhatkit==5.2
python-jose==3.3.0
python-multipart==0.0.6
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
